requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argparse==1.4.0
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
cryptography==46.0.7
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
future==1.0.0
gevent==25.9.1
grobid-client-python==0.1.4
grobid2json==0.0.1
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-ollama==1.0.1
langchain-openai==1.1.12
langchain-qdrant==1.1.0
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
magika==0.6.3
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.5
markupsafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy-extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
onnxruntime==1.20.1
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pdfminer-six==20251230
pdfplumber==0.11.9
pillow==12.2.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydotenv==0.0.7
pygments==2.19.2
pyjwt==2.12.1
pymongo==4.16.0
pypdfium2==5.7.0
pysocks==1.7.1
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
sentence-transformers==3.3.1
qdrant-client==1.17.1
regex==2026.3.32
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.56.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tldextract==5.3.1
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zope-event==6.1
zope-interface==8.2
zstandard==0.25.0
